Hybrid mode runs gumtree if correlation fails
GNU Diff is used to compute an expected fault position.
If that fails, Gumtree is used. The expected fault position
is used to perform a correlation.

If that correlation fails and the position was computed with GNU Diff,
a new correlation is performed with a Gumtree position.

Correlation file is cleaned up according to Gumtree knownledge.